openmcp-loadbalancing-controller/example/IoT-Data-Collector/docker/requirements.txt
beautifulsoup4==4.7.1
CoAPthon==4.0.2
paho-mqtt==1.4.0
requests==2.21.0
kafka==1.3.5
